expand playarea for scroll bar
When a card sticks off the bottom of the play area
(because a run of two or more cards is dragged near the bottom,
or because a pile has grown very tall, or because the window is resized to be short),
a vertical scroll bar appears on the play area.

On a Mac this poses no problem, but on Windows (and presumably on some other systems)
this scroll bar takes up some horizontal space, causing the play area to reflow
and one of the piles to head south for the winter.

On Windows we can avoid this reflow by reserving 15px of extra space to put the scroll bar in.